Now deals with auth such that if upstream link is closed it will re-auth with the current auth details

+++ ppp.c
@@ -54,30 +54,34 @@ static void ppp_timer_cb(evutil_socket_t fd, short what, void *arg)
 	PPPSession *pppSession=cp->pppSession;
 	if (cp==&pppSession->lcp) {
 		int next_state = pppSession->ppp.lcp;
-		LOG(3, pppSession->pppoeSession, "Got a timeout event for LCP\n");
-		switch (pppSession->ppp.lcp)
-		{
-		case RequestSent:
-		case AckReceived:
-			next_state = RequestSent;
-
-		case AckSent:
-			if (pppSession->lcp.conf_sent < ppp_max_configure)
+		LOG(3, pppSession->pppoeSession, "LCP: timeout: state %s, phase %s\n", ppp_state(pppSession->ppp.lcp), ppp_phase(pppSession->ppp.phase));
+		if (pppSession->ppp.phase==Authenticate) {
+			do_auth(pppSession);
+		} else {
+			switch (pppSession->ppp.lcp)
 			{
-				LOG(3, pppSession->pppoeSession, "No ACK for LCP ConfigReq... resending\n");
-				sendLCPConfigReq(pppSession);
-				change_state(pppSession, lcp, next_state);
-			}
-			else
-			{
-				sessionshutdown(pppSession, 1, "No response to LCP ConfigReq.");
-			}
-			break;
+			case RequestSent:
+			case AckReceived:
+				next_state = RequestSent;
 
-		case Closing:
-			LOG(3, pppSession->pppoeSession, "Timer expired on close - kill the session\n");
-			sessionkill(pppSession);
-			break;
+			case AckSent:
+				if (pppSession->lcp.conf_sent < ppp_max_configure)
+				{
+					LOG(3, pppSession->pppoeSession, "No ACK for LCP ConfigReq... resending\n");
+					sendLCPConfigReq(pppSession);
+					change_state(pppSession, lcp, next_state);
+				}
+				else
+				{
+					sessionshutdown(pppSession, 1, "No response to LCP ConfigReq.");
+				}
+				break;
+
+			case Closing:
+				LOG(3, pppSession->pppoeSession, "Timer expired on close - kill the session\n");
+				sessionkill(pppSession);
+				break;
+			}
 		}
 
 	} else if (cp==&pppSession->ipcp) {
@@ -512,7 +516,7 @@ void sessionshutdown(PPPSession *pppSession, int initiate, char const *reason)
 
 	start_close_timer(pppSession, lcp);
 
-	(*pppSession->cb)(pppSession, 4);
+	(*pppSession->cb)(pppSession, PPPCBACT_SHUTDOWN);
 }
 
 void sessionkill(PPPSession *pppSession)
